refactor(ui): move keymap helpers to dbflux_ui_base (#102)
Moves key_chord_from_gpui, modifiers_from_gpui, and default_keymap (with
all 15 layer builder fns) to dbflux_ui_base::keymap. Updates
dbflux_ui::keymap::mod.rs to re-export them from the new location.
Deletes the now-empty chord_ext.rs and defaults.rs source files.

All ~12 call-sites in dbflux_ui using crate::keymap::{default_keymap,
key_chord_from_gpui} are unchanged.
