+++ crates/dbflux_ui_base/src/lib.rs
@@ -5,7 +5,9 @@
 //! full `dbflux_ui` crate.
 
 pub mod app_state_entity;
+pub mod keymap;
 
 #[cfg(feature = "mcp")]
 pub use app_state_entity::McpRuntimeEventRaised;
 pub use app_state_entity::{AppStateChanged, AppStateEntity, AuthProfileCreated};
+pub use keymap::{default_keymap, key_chord_from_gpui};
